Refactor helper methods into RepositoryHelpers
