+++ test/UnitTests/UnitTests.csproj
@@ -266,6 +266,7 @@
     <Compile Include="Helpers\CommandTestHelpers.cs" />
     <Compile Include="Helpers\LazySubstitute.cs" />
     <Compile Include="Helpers\ReactiveTestHelper.cs" />
+    <Compile Include="Helpers\RepositoryHelpers.cs" />
     <Compile Include="Helpers\TestBaseClass.cs" />
     <Compile Include="Helpers\TestSharedCache.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
